moved saved variant selectors to own file
